move ogcapi-features config
