+++ src/extensions/ogcapi/features/src/main/resources/META-INF/spring.factories
@@ -1,3 +1,4 @@
 org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
 org.geoserver.cloud.autoconfigure.extensions.ogcapi.features.OgcApiFeaturesAutoConfiguration,\
+org.geoserver.cloud.autoconfigure.extensions.ogcapi.features.OgcApiFeaturesWebUIAutoConfiguration,\
 org.geoserver.cloud.autoconfigure.extensions.ogcapi.features.OgcApiFeaturesConformancesModuleAutoConfiguration
