+++ src/extensions/ogcapi/features/src/main/resources/META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports
@@ -1,2 +1,3 @@
 org.geoserver.cloud.autoconfigure.extensions.ogcapi.features.OgcApiFeaturesAutoConfiguration
+org.geoserver.cloud.autoconfigure.extensions.ogcapi.features.OgcApiFeaturesWebUIAutoConfiguration
 org.geoserver.cloud.autoconfigure.extensions.ogcapi.features.OgcApiFeaturesConformancesModuleAutoConfiguration
